Refactor: 포스트 레포지토리 구현
